+++ src/main/scala/io/viash/ViashNamespace.scala
@@ -114,7 +114,8 @@ object ViashNamespace extends Logging {
     append: Boolean = false,
     cpus: Option[Int],
     memory: Option[String],
-    dryRun: Option[Boolean] = None
+    dryRun: Option[Boolean] = None,
+    deterministicBuildFolder: Option[String] = None
   ): List[Either[(Config, ManyTestOutput), Status]] = {
     val configs1 = configs.filter{tup => tup match {
       // remove nextflow because unit testing nextflow modules
@@ -135,7 +136,11 @@ object ViashNamespace extends Logging {
     val tsvExists = tsvPath.exists(Files.exists(_))
     val tsvWriter = tsvPath.map(Files.newBufferedWriter(_, StandardOpenOption.CREATE, StandardOpenOption.WRITE, StandardOpenOption.APPEND))
 
-    val parentTempPath = IO.makeTemp("viash_ns_test")
+    val parentTempPath = IO.makeTemp(
+      name = deterministicBuildFolder.getOrElse("viash_ns_test"),
+      parentTempPath = None,
+      addRandomized = deterministicBuildFolder.isEmpty
+    )
     if (keepFiles.getOrElse(true)) {
       info(s"The working directory for the namespace tests is ${parentTempPath.toString()}")
     }