+++ src/main/scala/io/viash/ViashTest.scala
@@ -68,10 +68,15 @@ object ViashTest extends Logging {
     parentTempPath: Option[Path] = None, 
     cpus: Option[Int], 
     memory: Option[String],
-    dryRun: Option[Boolean] = None
+    dryRun: Option[Boolean] = None,
+    deterministicBuildFolder: Option[String] = None,
   ): ManyTestOutput = {
     // create temporary directory
-    val dir = IO.makeTemp("viash_test_" + config.functionality.name, parentTempPath)
+    val dir = IO.makeTemp(
+      name = deterministicBuildFolder.getOrElse(s"viash_test_${config.functionality.name}_"),
+      parentTempPath = parentTempPath,
+      addRandomized = deterministicBuildFolder.isEmpty
+    )
     if (!quiet) infoOut(s"Running tests in temporary directory: '$dir'")
 
     // set version to temporary value